Feat(FrontEnd/BackEnd): Melhorando mapa de calor e adicionando campo de numero de veiculos na tela

+++ WebAPI/src/main/java/br/com/crescer/monitorveiculos/modelo/CalculoEnergiaModel.java
@@ -21,5 +21,5 @@ public class CalculoEnergiaModel {
 
     private Double energia;
     
-    private Double numeroVeiculos;
+    private Long numeroVeiculos;
 }
